make database drop a config option

+++ test/lib/Suite.scala
@@ -36,7 +36,7 @@ trait WorkflowIntegrationSuite extends Suite with OneServerPerSuite with BeforeA
   }
 
   override def afterAll() {
-    DatabaseManager.destroy
+    if(Config.dropDB) DatabaseManager.destroy
   }
 
   override def beforeEach() {
